Improve DOOCS channel naming convention with scalar values

+++ liso/experiment/tests/test_machine.py
@@ -15,8 +15,8 @@ from . import DoocsDataGenerator as ddgen
 class TestDoocsMachine(unittest.TestCase):
     def setUp(self):
         m = EuXFELInterface(delay=0.01)
-        m.add_control_channel(dc.FLOAT32, "A/B/C/D")
-        m.add_control_channel(dc.FLOAT, "A/B/C/E")
+        m.add_control_channel(dc.FLOAT, "A/B/C/D")
+        m.add_control_channel(dc.DOUBLE, "A/B/C/E")
         m.add_instrument_channel(dc.IMAGE, "H/I/J/K", shape=(4, 4), dtype="uint16")
         m.add_instrument_channel(dc.IMAGE, "H/I/J/L", shape=(5, 6), dtype="float32")
         self._machine = m
